+++ packages/agent-toolkit/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @clerk/agent-toolkit
 
+## 0.1.36
+
+### Patch Changes
+
+- Updated dependencies [[`ea2bc26`](https://github.com/clerk/javascript/commit/ea2bc260fadac8fd7480cd476046f5a06c0d917d), [`37028ca`](https://github.com/clerk/javascript/commit/37028caad59cb0081ac74e70a44e4a419082a999)]:
+  - @clerk/backend@2.17.0
+  - @clerk/types@4.90.0
+  - @clerk/shared@3.27.1
+
 ## 0.1.35
 
 ### Patch Changes